requirements.txt
# FastAPI + LangChain + Supabase Backend

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12

# LangChain core
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-google-genai==2.0.0
langchain-postgres==0.0.12

# Google AI
google-generativeai>=0.7.0,<0.8.0   # fixed to match langchain-google-genai requirements

# Supabase / PostgreSQL
supabase==2.10.0
psycopg2-binary==2.9.9
pgvector>=0.2.5,<0.3.0             # fixed to match langchain-postgres requirements

# Document parsing
pypdf==5.0.0
python-docx==1.1.2
python-pptx==1.0.0
chardet==5.2.0

# URL crawling
httpx==0.27.0
beautifulsoup4==4.12.3
markdownify==0.13.1

# Utilities
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1